+++ app/src/utils/local_date/getLocalDate.test.ts
@@ -25,14 +25,9 @@ describe("Test get local date utility", () => {
         expect(data).toMatchObject(defaultResult)
     })
 
-    it("should return user formated date parts", () => {
-        const data = getLocalDate("Mon Sep 29 2025 10:14:09 GMT+0530 (India Standard Time)", {
-            weekday: 'short',
-            month: 'narrow',
-            day: "2-digit",
-            year: '2-digit'
-        })
-        expect(data).toMatchObject(userResult)
+    it("should return user formated date.week part", () => {
+        const data = getLocalDate("Mon Sep 29 2025 10:14:09 GMT+0530 (India Standard Time)", { weekday: 'short' })
+        expect(data.weekday).toEqual(userResult.weekday)
     })
 
     it("should return today formated date parts", () => {